Déplace `POST /api/utilisateur` dans les routes explicitement publiques
